Рефакторинг(добавление отдельного блока utils и чистка неиспользуемых файлов и кода)
